Fix bug with non-available Fx repo causing crash (now fails gracefully); Fixed issue with deleting widgets (List.remove uses location unless the argument is an object, so now we cast to Integer to force the usage of the overload)

+++ src/nitezh/ministock/domain/AndroidWidgetRepository.java
@@ -108,7 +108,7 @@ public class AndroidWidgetRepository implements WidgetRepository {
         if (!ids.contains(oldId)) {
             return;
         }
-        ids.remove(oldId);
+        ids.remove((Integer) oldId);  // Need to cast otherwise 'remove' uses location
         this.setIds(ids);
     }
